Fuzz the parsers, fix what that found, and run the tests under sanitizers
Implements Phase 2 of Documentation/review/2026-07-update-plan.md, plus the
Phase 3, 5 and 6 items that could be verified here.

Verified on Ubuntu 24.04 LTS with Qt 6.4.2: the normal build passes 32/32
tests, an ASan+UBSan build passes 32/32 with no sanitizer reports, and the
five fuzz targets run clean (liquivision 10.6M executions, cochran 73.7M,
ostctools 14.2M, parse-xml 3.0M, datatrak 2.0M).

Fuzzing infrastructure:

- tests/fuzz/ adds libFuzzer targets for the XML parser and the liquivision,
  datatrak, cochran and ostctools importers, behind SUBSURFACE_FUZZ_BUILD.
  These are the routines that read bytes somebody else produced.
- linux-fuzz-asan.yml runs a short smoke fuzz on pushes and pull requests and
  a longer sweep nightly, and runs the test suite under ASan and UBSan.
  SUBSURFACE_ASAN_BUILD has existed for years but was never switched on
  anywhere, so the suite had never been run under a sanitizer.

Defects these found, none of which were visible by reading:

- datatrak divided by the cylinder size without checking it, and that field is
  routinely zero, so the process died with SIGFPE. Found within seconds of the
  first fuzz run.
- Merging dives read past the end of the cylinder mapping array. The array is
  as long as the source dive's cylinder list, but the gas change events,
  sample sensor indices and tank sensor mappings that index it come from the
  log and need not refer to a cylinder the dive has. Sizes are now carried
  alongside the arrays through merge_events, merge_samples,
  merge_tank_sensor_mappings and the renumber helpers.
  dives/merged-tank-sensors.xml changes by one line as a result: the fixture
  contains a dive with one cylinder and a gas change to cylinder 2, so the
  recorded output was whatever the out of bounds read happened to return. An
  index that cannot be mapped now resolves to the dive's first cylinder.
  Whether such an event would be better dropped is a dive data question.
- import-suunto cast unaligned positions inside a database blob to float * and
  dereferenced them, which UBSan reports on x86 and which faults outright on
  stricter architectures. Now read with memcpy.
- mainwindow kept a unique_ptr<MapWidget> whether or not mapwidget.cpp was
  compiled, so a build without QtLocation only linked because an optimised
  build elided the references.

Liquivision, which carried the worst of the reviewed defects:

- The entry point added a file supplied length to the cursor unchecked, so a
  length larger than the file made the remaining size underflow to nearly 4GB
  and every bounds check further in became meaningless.
- The location and place strings were built directly from two 32 bit lengths
  out of the file.
- The sample count check multiplied before comparing, so a count of 0x40000000
  wrapped to zero and passed the check meant to stop it.
- The event walker read up to 26 bytes per event with nothing keeping it inside
  the buffer.
- core/bounded-reader.h adds the checked cursor used for this, so the same
  shape does not have to be rewritten by hand for the next importer.

Other untrusted input handling:

- Drop XML_PARSE_HUGE where the input is a file the user opened or a server
  response; it disables libxml2's depth limit while traverse()/visit() recurse
  over the tree. Add XML_PARSE_NONET everywhere.
- Deny XSLT transforms the ability to write files, create directories or use
  the network. Reading is left alone deliberately: the stylesheets pull in
  commonTemplates.xsl with xsl:include, and the loader already confines reads
  to the ":/xslt" resource.
- Cap zip decompression and move the growth arithmetic off int, where it
  overflowed to a negative size past about 1.4GB. Close the file descriptor
  readfile() leaked on every call, stop skipping zip_fclose() for picture
  entries, and NULL check zip_get_name().

Concurrency and correctness:

- import_thread_cancelled is written from the GUI thread and polled by the
  download thread; as a plain int that is a race and the compiler was free to
  hoist the load out of the loop.
- The planner held a mutex by hand across a couple of hundred lines that can
  throw, so an exception left it locked for the rest of the session.
- unregister_dive() takes size_t now rather than relying on npos narrowing to
  -1, and removeDive() stops instead of storing a null dive for undo to use.

Supply chain, hardening and docs:

- get-dep-lib.sh fetched libftdi over git://, which has no TLS; it now uses the
  same host over https, verifies downloads against checksums.sha256 when a
  checksum is on record, and passes --fail so an HTTP error is not saved and
  unpacked as if it were a tarball.
- Enable _FORTIFY_SOURCE, -fstack-protector-strong and full RELRO.
- Add SECURITY.md, which did not exist, describing where to report and what
  the interesting attack surface is.
- INSTALL.md was still titled "Getting Qt5" and quoted Qt 5.9/5.12 floors while
  mobile requires KF6 and Qt6; state the real supported versions. Note in
  README.md that the absence of tags since 2022 is the move to CI/CD releases
  rather than the project having stopped.

+++ core/divelist.h
@@ -25,7 +25,11 @@ struct dive_table : public sorted_owning_table<dive, &comp_dives> {
 	dive *get_by_uniq_id(int id) const;
 	void record_dive(std::unique_ptr<dive> d);	// call fixup_dive() before adding dive to table.
 	struct dive *register_dive(std::unique_ptr<dive> d);
-	std::unique_ptr<dive> unregister_dive(int idx);
+	// Takes the same size_t index type that get_idx() returns. It used to take
+	// an int, so a not-found result (std::string::npos) only behaved because
+	// narrowing turned it into -1 and the implementation happened to test for
+	// a negative index.
+	std::unique_ptr<dive> unregister_dive(size_t idx);
 	std::unique_ptr<dive> default_dive();		// generate a sensible looking defaultdive 1h from now.
 
 	// Some of these functions act on dives, but need data from adjacent dives,
